scripts/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
python-dotenv>=1.0.0

# Supabase
supabase>=2.0.0

# GPX Generation
gpxpy>=1.6.0

# Utilities
tqdm>=4.66.0
